git-svn-id: https://clonekeenplus.svn.sourceforge.net/svnroot/clonekeenplus/cgenius/trunk@905 4df4b0f3-56ce-47cb-b001-ed939b7d65a6

+++ src/sdl/CInput.cpp
@@ -126,7 +126,7 @@ bool CInput::startJoyDriver()
 	else
 	{
 		int i=0;
-		int joynum;
+		size_t joynum;
 		joynum = SDL_NumJoysticks();
 		if(joynum)
 		{
